backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
psycopg[binary]==3.3.2

alembic==1.14.0
pydantic-settings==2.6.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
